fixSidebar
