+++ src/components/Sidebar/Sidebar.css
@@ -1,0 +1,75 @@
+.sidebar-sider {
+    overflow: auto;
+    height: 100vh;
+    position: sticky;
+    top: 0;
+    left: 0;
+    box-shadow: 2px 0 8px rgba(0, 0, 0, 0.1);
+  }
+  
+  .sidebar-logo {
+    height: 64px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 16px 0;
+  }
+  
+  .sidebar-logo img {
+    margin-right: 10px;
+    height: 40px;
+    width: 40px;
+    border-radius: 50%;
+    object-fit: cover;
+  }
+  
+  .sidebar-menu {
+    border-right: 0;
+  }
+  
+  .sidebar-notification {
+    padding: 16px;
+    text-align: center;
+    color: #f5222d;
+    position: absolute;
+    bottom: 20px;
+    width: 100%;
+    background: #fff;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+    border-radius: 8px;
+    overflow: hidden;
+    animation: fadeInUp 0.5s ease;
+  }
+  
+  .sidebar-notification.hidden {
+    display: none;
+  }
+  
+  .notify-bold {
+    margin: 0;
+    font-weight: bold;
+  }
+  
+  .notify-location {
+    margin: 0;
+  }
+  
+  .shake {
+    display: inline-block;
+    animation: shake 1s infinite;
+    color: #faad14;
+    font-size: 24px;
+    margin-bottom: 8px;
+  }
+  
+  @keyframes shake {
+    0%, 100% { transform: translateX(0); }
+    20%, 60% { transform: translateX(-5px); }
+    40%, 80% { transform: translateX(5px); }
+  }
+  
+  @keyframes fadeInUp {
+    from { opacity: 0; transform: translateY(20px); }
+    to { opacity: 1; transform: translateY(0); }
+  }
+
